Avoid importing __version__ in setup.py
Instead of importing __version__, parse qnet/__init__.py for the version
string. The way __init__.py is currently written, it doesn't make a
difference, but if we were ever to change the file such that the import
may fail due to unmet dependencies (e.g., by exposing any of
the subpackages), it would render qnet uninstallable.

See:
https://www.python.org/dev/peps/pep-0396/
http://stackoverflow.com/questions/2058802/how-can-i-get-the-version-defined-in-setup-py-setuptools-in-my-package