Set automatic filename in add_observable method
Instead of passing a filename to the add_observable method, we only pass
a name, and create a filename based on that name. This ensures that we
can safely make some assumptions, such as that the filename is not a
full path, so we can safely create in inside any temporary folder.