Give Feedback I/O ports as kwargs, not operands
Whether the Feedback port numbers are "operands" (which should be
"members of the algebra") is at the very least questionable.
Furthermore, by making out_port and in_port keyword-only arguments, we
make the code more explicit and avoid bugs due to mixing up input and
output. Lastly, renaming the parameters in in preparation for #9