Add 'identifier' property to LocalOperators
Move the handling of custom identifiers from the printer into a
property. This elegantly allows to tie the Create and Destroy
identifiers together. The alternative plan formulated in #60 to get rid
of the Create class turned out to be a dead end: Formulating rules
becomes very cumbersome without an explicit Create and Destroy class.