Test equality and hashing of neutral elements
With the implementation of the scalar algebra it turns out that the way
to go is to have only the scalar Zero and One equal to 0 and 1. In every
other algebra, the neutral elements have no relation to the scalar 0 and
1, or to the neutral elements of other algebras.

Hoever, elements of "quantum" algebras (anything that has a Hilbert
space, including in fact the scalar algebra on the TrivialSpace) have an
is_zero attribute to compare to the abstract zero.

This closes #63