Remove ABCD class
The ABCD class was not tested, and its implementation appears
incomplete.  Nobody has come forward to give any input on this, or to
maintain it in any way, so I'm removing it. There is still the getABCD
function that seems to contain all the non-trivial code related to
linearized models.

This code could be re-added if someone needs it and is available to
complete the implementation and to add tests and documentation.

Closes #72