Fix collection of terms in sums
There is really no way to have an ordering that guarantees that two
terms a * A, b * A in a sum are next to each other, so that the
rule -> (a + B) * A can be applied reliably. The solution is to have
have a "simplification" that iterates over all summands, instead of just
looking at next-neighbors.